requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
google-api-core==2.19.1
google-auth==2.31.0
google-geo-type==0.3.7
google-maps-places==0.1.15
googleapis-common-protos==1.63.2
googlemaps==4.10.0
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
idna==3.7
Markdown==3.6
packaging==24.1
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dotenv==1.0.1
requests==2.32.3
requests-mock==1.12.1
rsa==4.9
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
